+++ lib/pay/stripe.rb
@@ -3,6 +3,8 @@ module Pay
     class Error < Pay::Error
     end
 
+    autoload :Sync, "pay/stripe/sync"
+
     module Webhooks
       autoload :AccountUpdated, "pay/stripe/webhooks/account_updated"
       autoload :ChargeRefunded, "pay/stripe/webhooks/charge_refunded"